// client/src/api/model/initiativeSlotResults.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export interface InitiativeSlotResults {
  /** Number of Success symbols rolled */
  success: number;
  /** Number of Advantage symbols rolled */
  advantage: number;
  /** Number of Triumph symbols rolled */
  triumph: number;
  /** Number of Failure symbols rolled */
  failure: number;
  /** Number of Threat symbols rolled */
  threat: number;
  /** Number of Despair symbols rolled */
  despair: number;
}